Generalize mRNA artifacts and configuration
