+++ src/main/groovy/org/transmartproject/batch/beans/AbstractJobConfiguration.groovy
@@ -1,31 +1,10 @@
 package org.transmartproject.batch.beans
 
-import groovy.transform.TypeChecked
-import org.codehaus.groovy.runtime.MethodClosure
 import org.springframework.batch.core.Job
 import org.springframework.batch.core.JobExecutionListener
 import org.springframework.batch.core.Step
-import org.springframework.batch.core.StepExecutionListener
 import org.springframework.batch.core.configuration.annotation.JobBuilderFactory
 import org.springframework.batch.core.configuration.annotation.JobScope
-import org.springframework.batch.core.configuration.annotation.StepBuilderFactory
-import org.springframework.batch.core.configuration.annotation.StepScope
-import org.springframework.batch.core.job.builder.FlowBuilder
-import org.springframework.batch.core.job.flow.Flow
-import org.springframework.batch.core.job.flow.support.SimpleFlow
-import org.springframework.batch.core.step.tasklet.Tasklet
-import org.springframework.batch.item.ItemProcessor
-import org.springframework.batch.item.ItemStreamReader
-import org.springframework.batch.item.file.FlatFileItemReader
-import org.springframework.batch.item.file.LineCallbackHandler
-import org.springframework.batch.item.file.mapping.BeanWrapperFieldSetMapper
-import org.springframework.batch.item.file.mapping.DefaultLineMapper
-import org.springframework.batch.item.file.mapping.FieldSetMapper
-import org.springframework.batch.item.file.mapping.PassThroughFieldSetMapper
-import org.springframework.batch.item.file.separator.DefaultRecordSeparatorPolicy
-import org.springframework.batch.item.file.transform.DelimitedLineTokenizer
-import org.springframework.batch.item.support.CompositeItemProcessor
-import org.springframework.batch.item.validator.Validator
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.config.BeanFactoryPostProcessor
 import org.springframework.beans.factory.config.ConfigurableListableBeanFactory
@@ -34,21 +13,16 @@ import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.ComponentScan
 import org.springframework.context.annotation.Import
 import org.springframework.context.support.ConversionServiceFactoryBean
-import org.springframework.context.support.ResourceBundleMessageSource
 import org.springframework.core.convert.converter.Converter
-import org.springframework.core.io.Resource
 import org.springframework.jdbc.core.JdbcTemplate
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate
-import org.transmartproject.batch.batchartifacts.*
-import org.transmartproject.batch.clinical.facts.ClinicalDataRow
+import org.transmartproject.batch.batchartifacts.BetterExitMessageJobExecutionListener
+import org.transmartproject.batch.batchartifacts.DefaultJobIncrementer
 import org.transmartproject.batch.db.*
-import org.transmartproject.batch.facts.ClinicalFactsRowSet
 
 import javax.sql.DataSource
 import java.nio.file.Path
 import java.nio.file.Paths
-
-import static org.springframework.batch.item.file.transform.DelimitedLineTokenizer.DELIMITER_TAB
 
 /**
  * Base class for Spring context configuration classes for Jobs.
@@ -59,7 +33,7 @@ import static org.springframework.batch.item.file.transform.DelimitedLineTokeniz
         'org.transmartproject.batch.db',
         'org.transmartproject.batch.support',
 ])
-abstract class AbstractJobConfiguration {
+abstract class AbstractJobConfiguration implements StepBuildingConfigurationTrait {
 
     public static final int DEFAULT_FETCH_SIZE = 1000
 
@@ -67,7 +41,7 @@ abstract class AbstractJobConfiguration {
     JobBuilderFactory jobs
 
     @Autowired
-    StepBuilderFactory steps
+    MessageSource validationMessageSource
 
     abstract Job job()
 
@@ -123,174 +97,6 @@ abstract class AbstractJobConfiguration {
     protected void configure(SequenceReserver sequenceReserver) {
         sequenceReserver.defaultBlockSize = 10
     }
-
-    final protected Step stepOf(String name, Tasklet tasklet) {
-        steps.get(name)
-                .tasklet(tasklet)
-                .listener(showCountStepListener())
-                .build()
-    }
-
-    final protected Step allowStartStepOf(String name, Tasklet tasklet) {
-        steps.get(name)
-                .allowStartIfComplete(true)
-                .tasklet(tasklet)
-                .listener(showCountStepListener())
-                .build()
-    }
-
-    final protected Step stepOf(MethodClosure closure) {
-        stepOf(closure.method - ~/^get/, closure.call())
-    }
-
-    final protected Step allowStartStepOf(MethodClosure closure) {
-        allowStartStepOf(closure.method - ~/^get/, closure.call())
-    }
-
-    @Bean
-    StepExecutionListener showCountStepListener() {
-        new LogCountsStepListener()
-    }
-
-    Flow flowOf(Step step) {
-        new FlowBuilder<SimpleFlow>().start(step).build()
-    }
-
-    Flow parallelFlowOf(String name, Step step, Step... otherSteps) {
-        new FlowBuilder<SimpleFlow>(name)
-                .start(step)
-                //forks execution
-                //need to use a tweaked executor. see https://jira.spring.io/browse/BATCH-2269
-                .split(new JobContextAwareTaskExecutor())
-                .add(otherSteps.collect { flowOf(it) } as Flow[])
-                .end()
-    }
-
-    ItemProcessor compositeOf(ItemProcessor... processors) {
-        CompositeItemProcessor<ClinicalDataRow, ClinicalFactsRowSet> result =
-                new CompositeItemProcessor<ClinicalDataRow, ClinicalFactsRowSet>()
-        result.setDelegates(processors.toList())
-        result
-    }
-
-    @Bean
-    @StepScope
-    HeaderSavingLineCallbackHandler headerSavingLineCallbackHandler() {
-        new HeaderSavingLineCallbackHandler()
-    }
-
-    @Bean
-    @StepScope
-    ProgressWriteListener progressWriteListener() {
-        new ProgressWriteListener()
-    }
-
-    @Bean
-    @StepScope
-    LineOfErrorDetectionListener lineOfErrorDetectionListener() {
-        new LineOfErrorDetectionListener()
-    }
-
-    @TypeChecked
-    protected
-    <T> ItemStreamReader<T> tsvFileReader(Map<String, Object> options,
-                                    Resource resource) {
-        /* options:
-         * strict: true|false (default: yes)
-         * saveState: true|false (default: yes)
-         * beanClass: JavaBean class
-         *            (optional, o/wise use PassThroughFieldSetMapper)
-         * columnNames: names of columns, required if mapping into bean
-         *              properties (List<String> or String[])). Can be
-         *              'auto' to read names from header (linesToSkip
-         *              must be 1 then)
-         * linesToSkip: <int> (default: 0)
-         * saveHeader: true|false|LineCallbackHandler object (default: false)
-         *
-         */
-        def strict = options.containsKey('strict') ?
-                (boolean) options.strict : true
-        def saveHeader = options.containsKey('saveHeader') ?
-                options.saveHeader : false
-        int linesToSkip = options.containsKey('linesToSkip') ?
-                (options.linesToSkip as int) : 0
-
-        LineCallbackHandler lch = null
-        if (saveHeader) {
-            if (saveHeader instanceof LineCallbackHandler) {
-                lch = (LineCallbackHandler) saveHeader
-            } else {
-                lch = headerSavingLineCallbackHandler()
-            }
-        }
-
-        FieldSetMapper<T> mapper
-        if (options.beanClass) {
-            mapper = new BeanWrapperFieldSetMapper<T>(
-                    targetType: (Class<? extends T>) options.beanClass)
-        } else {
-            mapper = (FieldSetMapper<T>) new PassThroughFieldSetMapper()
-        }
-
-        DelimitedLineTokenizer tokenizer =  new DelimitedLineTokenizer(
-                names: ((options.columnNames && options.columnNames != 'auto') ?
-                        options.columnNames : []) as String[],
-                delimiter: DELIMITER_TAB,
-        )
-
-        /* if columnNames is 'auto', read the first line to determine
-         * the identity of the columns */
-        if (options.columnNames == 'auto') {
-            def originalLch = lch
-            lch = { String line ->
-                if (originalLch) {
-                    originalLch.handleLine line
-                }
-                tokenizer.names = new DelimitedLineTokenizer(DELIMITER_TAB)
-                        .tokenize(line).values
-            }
-        }
-
-        if (linesToSkip == 0 && lch) {
-            throw new IllegalArgumentException(
-                    'Cannot look at header is there are no lines to skip')
-        }
-
-        new FlatFileItemReader<T>(
-                lineMapper: new DefaultLineMapper<T>(
-                        lineTokenizer: tokenizer,
-                        fieldSetMapper: mapper,
-                ),
-
-                linesToSkip: linesToSkip,
-                skippedLinesCallback: lch,
-
-                recordSeparatorPolicy: new DefaultRecordSeparatorPolicy(),
-                resource: resource,
-                strict: strict,
-                saveState: options.containsKey('saveState') ?
-                        (boolean) options.saveState : true,
-        )
-    }
-
-    @Bean
-    MessageSource validationMessageSource() {
-        new ResourceBundleMessageSource(
-                basename: 'org.transmartproject.batch.validation_messages',
-                defaultEncoding: 'UTF-8')
-    }
-
-
-    static protected final Step wrapStepWithName(final String name,
-                                                 final Step step) {
-        /* bad hack to avoid getName() being called on a scoped proxy */
-        new OverriddenNameStep(step: step, newName: name,)
-    }
-
-    protected Validator adaptValidator(org.springframework.validation.Validator springValidator) {
-        new MessageResolverSpringValidator(springValidator, validationMessageSource())
-    }
-
 }
 
 /* needed so the runtime can know the generic types */
