+++ src/main/groovy/org/transmartproject/batch/highdim/datastd/StandardDataValue.groovy
@@ -1,15 +1,15 @@
-package org.transmartproject.batch.highdim.mrna.data.pass
+package org.transmartproject.batch.highdim.datastd
 
 import groovy.transform.Canonical
 import org.transmartproject.batch.highdim.compute.DataPoint
 import org.transmartproject.batch.patient.Patient
 
 /**
- * Single value for a (patient, probe) pair.
+ * Single value for a (patient, annotation) pair.
  */
 @Canonical
-class MrnaDataValue implements DataPoint {
-    String probe
+class StandardDataValue implements DataPoint {
+    String annotation
     Patient patient
     Double value
 }
